store/pyproject.toml
[tool.poetry]
name = "store"
version = "0.1.0"
description = ""
authors = ["MattheusBy <mattaudia6@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.5"
celery = "^5.2.7"
bs4 = "^0.0.1"
django-rest-framework = "^0.1.0"
dj-database-url = "^1.2.0"
django-simple-captcha = "^0.5.17"
django-paypal = "^2.0"
django-celery-beat = "^2.4.0"
django-googledrive-storage = "^1.6.0"
django-widget-tweaks = "^1.4.12"
psycopg2-binary = "^2.9.5"
lxml = "^4.9.2"
django-cors-headers = "^3.13.0"
whitenoise = "^6.3.0"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
